+++ sidebars/sidebarsMemgraph.js
@@ -175,7 +175,7 @@ module.exports = {
           ],
         },
         "how-to-guides/networkx",
-        "how-to-guides/query-modules/overview",
+        "how-to-guides/query-modules",
         "how-to-guides/replication",
         //"how-to-guides/tensorflow-setup",
         "reference-guide/triggers",