+++ docs/changelog.md
@@ -3,6 +3,9 @@ id: changelog
 title: Changelog
 sidebar_label: Changelog
 ---
+## v2.3.0
+
+- Added foreach clause [#351](https://github.com/memgraph/memgraph/pull/351)
 
 ## v2.2.1 - Mar 17, 2022